pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drf-easy-crud"
version = "0.0.1"
description = "Easy CRUD operations for Django REST Framework."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Avi Zaguri" },]
keywords = ["development", "python", "django"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Framework :: Pytest",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "custom-python-logger>=2.0.11",
    "pytest-plugins>=2.0.1",
    "python-base-toolkit>=0.0.17",
    "python-custom-exceptions>=0.0.2",
    "python-github-plus==0.0.5",
    "beautifulsoup4>=4.14.3",
    "pre-commit>=4.5.0",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "tenacity>=9.1.2",
    "wheel>=0.45.1",
    "django>=5.2.11",
    "djangorestframework>=3.16.1",
    "django-schema-graph>=3.1.0",
    "django-filter>=25.2",
    "pytest-django>=4.11.1",
    "django-models>=1.0",
]

[project.urls]
Homepage = "https://github.com/aviz92/drf-easy-crud"
Repository = "https://github.com/aviz92/drf-easy-crud"
Issues = "https://github.com/aviz92/drf-easy-crud/issues"

[tool.setuptools.packages.find]
where = ["."]
#include = ["drf_easy_crud*"]
namespaces = false
